jobs/embed-openalex/pyproject.toml
[tool.poetry]
name = "embed-openalex"
version = "0.1.0"
description = ""
authors = ["Antoine Gerardin <antoine.gerardin@nih.gov>"]
readme = "README.md"
packages = [{include = "polus", from = "src"}]


[tool.poetry.dependencies]
python = ">=3.10,<3.12"
polus-aithena-common = {path= "../../common", develop=true}
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
python-dotenv = "^1.0.1"
# TODO switch to git dependency when stable or better pypi
openalex-types = {path="/polus2/gerardinad/projects/openalex-types", develop=true}
tqdm = "^4.67.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
